GROK-13932: NLP | SBS: Refactoring
Refactored code for stemming-based search
