+++ packages/NLP/css/stemming-search.css
@@ -1,0 +1,4 @@
+.nlp-stemming-edit {
+    cursor: pointer;
+    color: #3cb173;
+}
