+++ packages/NLP/src/stemming-tools/stemming-tools.ts
@@ -8,14 +8,9 @@ import * as DG from 'datagrok-api/dg';
 
 import {stemmer} from 'stemmer';
 import {UMAP} from 'umap-js';
-import {STR_METRIC_TYPE, MetricInfo, getDefaultMetric,
-  getMetricTypesChoicesList, DISTANCE_TYPE, getSingleSourceMetricsMap, getDefaultDistnce, DIST_TYPES_ARR,
+import {STR_METRIC_TYPE, MetricInfo, DISTANCE_TYPE, getSingleSourceMetricsMap, getDefaultDistnce,
   getDistanceFn, getMetricFn} from './metrics';
-
-/** Error messeges. */
-enum ERR_MSG {
-  UNSUPPORTED_METRIC = 'Unsupported metric',
-};
+import {RATIO, INFTY, TINY, MIN_CHAR_COUNT, STOP_WORDS, SEPARATORS} from './constants';
 
 /** Word information: index in the dictionary and frequency. */
 type WordInfo = {
@@ -36,7 +31,7 @@ type StemCash = {
 };
 
 /** Column (feature) usage specification: type, metric & use. */
-type ColUseInfo = {
+export type ColUseInfo = {
   type: string,
   metric: MetricInfo,
   use: boolean,
@@ -54,46 +49,6 @@ export var stemCash: StemCash = {
   aggrDistance: undefined,
   filters: undefined,
 };
-
-const RATIO = 0.05;
-
-/** A big number. */
-const INFTY = 100000;
-
-/** A small number. */
-const TINY = 0.000001;
-
-/** Minimal number of characters in a word. */
-const MIN_CHAR_COUNT = 1;
-
-/** Words that are skipped, when applying stemming-based search. */
-const STOP_WORDS = ['a', 'about', 'above', 'after', 'again', 'against',
-  'all', 'am', 'an', 'and', 'any', 'are', 'as', 'at',
-  'be', 'because', 'been', 'before', 'being', 'below', 'between', 'both', 'but', 'by',
-  'could',
-  'did', 'do', 'does', 'doing', 'down', 'during',
-  'each',
-  'few', 'for', 'from', 'further',
-  'had', 'has', 'have', 'having', 'he', 'he\'d', 'he\'ll', 'he\'s', 'her', 'here',
-  'here\'s', 'hers', 'herself', 'him', 'himself', 'his', 'how', 'how\'s',
-  'i', 'i\'d', 'i\'ll', 'i\'m', 'i\'ve', 'if', 'in', 'into', 'is', 'it', 'it\'s', 'its', 'itself', 'i.e',
-  'let\'s',
-  'me', 'more', 'most', 'my', 'myself',
-  'nor', 'nan', 'no', 'not',
-  'of', 'on', 'once', 'only', 'or', 'other', 'ought', 'our', 'ours', 'ourselves', 'out', 'over', 'own',
-  'same', 'she', 'she\'d', 'she\'ll', 'she\'s', 'should', 'so', 'some', 'such',
-  'than', 'that', 'that\'s', 'the', 'their', 'theirs', 'them', 'themselves', 'then',
-  'there', 'there\'s', 'these', 'they', 'they\'d', 'they\'ll',
-  'they\'re', 'they\'ve', 'this', 'those', 'through', 'to', 'too',
-  'under', 'until', 'up',
-  'very',
-  'was', 'we', 'we\'d', 'we\'ll', 'we\'re', 'we\'ve', 'were',
-  'what', 'what\'s', 'when', 'when\'s', 'where', 'where\'s', 'which',
-  'while', 'who', 'who\'s', 'whom', 'why', 'why\'s', 'with', 'would',
-  'you', 'you\'d', 'you\'ll', 'you\'re', 'you\'ve', 'your', 'yours', 'yourself', 'yourselves'];
-
-/** Separator characters. */
-const SEPARATORS = '[].,:;?!(){} \n\t#';
 
 /** Feature operating specification. */
 type FeatureSpecification = {
@@ -346,122 +301,6 @@ export function getMarkedString(curIdx: number, queryIdx: number, strToBeMarked:
   return marked;
 } // getMarkedStringAndCommonWordsMap
 
-/** Modify similarity metric. */
-export function modifyMetric(df: DG.DataFrame): void {
-  const colsData = new Map<string, ColUseInfo>();
-  const colsInp = new Map<string, {metricInput: DG.InputBase, weightInput: DG.InputBase}>();
-  const inputElements = new Map<string, HTMLDivElement>();
-  const dlg = ui.dialog({title: 'Edit distance'});
-  const initCheckedCols = [...stemCash.metricDef!.keys()];
-
-  const onColumnsChanged = (columns: DG.Column[]) => {
-    const names = columns.map((col) => col.name);
-
-    distInput.root.hidden = (names.length < 2);
-    columnsHeader.hidden = (names.length < 1);
-
-    for (const key of colsData.keys()) {
-      const val = colsData.get(key);
-
-      if (names.includes(key)) {
-        val!.use = true;
-        //@ts-ignore
-        inputElements.get(key)?.hidden = false;
-      } else {
-        val!.use = false;
-        //@ts-ignore
-        inputElements.get(key)?.hidden = true;
-      }
-
-      colsData.set(key, val!);
-    }
-  };
-
-  const colsInput = ui.columnsInput('Features', df, onColumnsChanged, {checked: initCheckedCols});
-  colsInput.setTooltip('Features used in computing similarity measure.');
-  dlg.add(ui.h3('Source'));
-  dlg.add(colsInput);
-
-  const distInput = ui.choiceInput(
-    'Distance',
-    stemCash.aggrDistance,
-    DIST_TYPES_ARR,
-    (dist: DISTANCE_TYPE) => {stemCash.aggrDistance = dist;},
-  );
-
-  distInput.setTooltip('Type of distance between elements with the specified features.');
-  dlg.add(distInput);
-  distInput.root.hidden = (initCheckedCols.length < 2);
-
-  const columnsHeader = ui.h3('Features');
-  dlg.add(columnsHeader);
-  columnsHeader.hidden = (initCheckedCols.length < 1);
-
-  // add an appropriate inputs
-  for (const col of df.columns) {
-    const name = col.name;
-
-    const colData = {
-      type: col.type,
-      metric: getDefaultMetric(col),
-      use: false,
-    };
-
-    if (stemCash.metricDef?.has(name)) {
-      const val = stemCash.metricDef.get(name);
-      colData.use = true;
-      colData.metric.type = val!.type;
-      colData.metric.weight = val!.weight;
-    }
-
-    colsData.set(name, colData);
-
-    const choices = getMetricTypesChoicesList(col);
-    const metricInput = ui.choiceInput(`${name}:`, colData.metric.type as string, choices, (str: string) => {
-      const val = colsData.get(name);
-      //@ts-ignore
-      val?.metric.type = str;
-      colsData.set(name, val!);
-    });
-    metricInput.setTooltip(`Type of metric between the '${name}' feature values.`);
-
-    const weightInput = ui.floatInput('metric with the weight', colData.metric.weight, (w: number) => {
-      const val = colsData.get(name);
-      //@ts-ignore
-      val?.metric.weight = w;
-      colsData.set(name, val!);
-    });
-    weightInput.setTooltip(`Weight coefficient of the '${name}' feature metric.`);
-
-    const inputs = {metricInput: metricInput, weightInput: weightInput};
-
-    colsInp.set(name, inputs);
-
-    const uiElem = ui.divH([inputs.metricInput.root, inputs.weightInput.root]);
-    inputElements.set(name, uiElem);
-
-    uiElem.hidden = !colData.use;
-
-    dlg.add(uiElem);
-  } // for
-
-  dlg
-    .onCancel(() => {})
-    .onOK(() => {
-      const map = new Map<string, MetricInfo>();
-
-      for (const key of colsData.keys()) {
-        const val = colsData.get(key);
-
-        if (val?.use)
-          map.set(key, val.metric);
-      }
-
-      stemCash.metricDef = map;
-    })
-    .show({x: 300, y: 300});
-} // modifyMetric
-
 /** Set stemming cash data. */
 export function setStemmingCash(df: DG.DataFrame, source: DG.Column): void {
   if ((stemCash.dfName !== df.name) || (stemCash.colName !== source.name)) {
@@ -499,7 +338,7 @@ function getTextDistFn(metric: MetricInfo): (query: Uint16Array, current: Uint16
     };
 
   default:
-    throw new Error(ERR_MSG.UNSUPPORTED_METRIC);
+    throw new Error('Unsupported metric');
   }
 }
